[EDIT] missing arg

+++ db/query/workout.sql
@@ -24,4 +24,4 @@ lifts = IFNULL(sqlc.arg('lifts'), lifts)
 WHERE id = ? AND user_id = UUID_TO_BIN(sqlc.arg('user_id'));
 
 -- name: DeleteWorkout :execresult
-DELETE FROM lift WHERE id = ? AND user_id = UUID_TO_BIN(sqlc.arg(''));
+DELETE FROM lift WHERE id = ? AND user_id = UUID_TO_BIN(sqlc.arg('user_id'));
